Added a paused function to GameEngine.js, SpiderFish.js, Entity.js, HumanBullet.js, Explosion.js, Animation.js That allows the GameEngine.js rather then doing the standard update function switch to keeping time sensitive objects alive (HumanBullet.js, and animations). This is managed by the new this.paused variable that is declared in GameEngine.js.
Added MainMenu.js to display a menu at the start of the game, and whenever the key defined at STANDARD.MENU_CONTROL_KEY is clicked, at the end of the game it cleans up the current game object, and starts a new game.

Added a new variable in STANDARD.js called STANDARD.MENU_CONTROL_KEY this is used to check if the menu should display/hide and the game should be paused/unpaused.

Signed-off-by: StepheNu <nuskos@uwindsor.ca>